hubble: return LostEvent on invalid ring buffer reads
In theory, the ring buffer writer can write faster than a ring reader
may read. When this situation happens, the event to be read might
already have been overwritten by the writer and is thus lost for the
reader. With the previous code, such events were simply dismissed. This
commit changes the behavior to return LostEvents when a read-miss
happens.

Signed-off-by: Robin Hahling <robin.hahling@gw-computing.net>