helm: fix TLS cert server name for cluster names containing dots with certgen
Before this patch, when using Helm installing with
--set hubble.tls.auto.method=cronJob and a cluster name containing dots
(.), Relay would fail the mTLS handshake with Hubble.

The root cause is that we don't provide the cluster name info to certgen
through Helm, and certgen then assumes that the cluster name is
"default".

This PR ensure that the common name for the Hubble certificate is given
to certgen through Helm in a way that is consistent with the peer
service and also hubble.tls.auto.method=helm.

See also https://github.com/cilium/cilium/pull/14413

Fix https://github.com/cilium/cilium/issues/14392

Signed-off-by: Alexandre Perrin <alex@kaworu.ch>